+++ docs/changelog.md
@@ -6,6 +6,7 @@
 - Added mod-accessible api NpcAdventureModApi (thx aedenthorn aka iLoveLucy)
 - Loading of legacy content packs (format 1.2 and older) is now DISABLED by default. (User can enable support for legacy packs in config file)
 - Added support for companion events (for future NA cutscenes and for use in content packs)
+- Added support for custom weapons for all combat levels (supports JSON assets weapons)
 
 ## 0.14.1 "Major Minority"